pyproject.toml
[project]
name = "nemoclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "watchfiles>=1.0",
    "structlog>=24.0",
    "croniter>=1.3",
]
